test(chunkStreams): rename & update related tests
